MR-64- Clean Archeticture
