--- app/src/main/java/com/example/myriyal/screens/authentication/data/repository/BaseAuthenticator.kt
@@ -1,14 +1,0 @@
-package com.example.myriyal.screens.authentication.data.repository
-
-import com.google.firebase.auth.FirebaseUser
-
-interface BaseAuthenticator{
-
-    suspend fun signUpWithEmailPassword(
-        username:String,
-        email:String,
-        password:String
-    ):FirebaseUser?
-
-
-}
