+++ app/src/main/java/com/example/myriyal/screens/authentication/domain/repository/AuthRepo.kt
@@ -1,0 +1,13 @@
+package com.example.myriyal.screens.authentication.domain.repository
+
+
+import com.google.firebase.auth.FirebaseUser
+
+interface AuthRepo {
+    suspend fun signUpWithEmailPassword(
+        username: String,
+        email: String,
+        password: String
+    ): FirebaseUser?
+
+}
