--- app/src/main/java/com/example/myriyal/screens/authentication/domain/repository/BaseAuthRepository.kt
@@ -1,20 +1,0 @@
-package com.example.myriyal.screens.authentication.domain.repository
-
-import com.google.firebase.auth.FirebaseUser
-
-
-/*
-use this interface to implement all authentication functions
-[Sign up], [Log in], [Log out], [Reset password]
-all classes that implement those functionality will be inherit from this interface!
- */
-
-
-interface BaseAuthRepository{
-
-    suspend fun signUpWithEmailPassword(
-        username:String,
-        email: String,
-        password: String) : FirebaseUser?
-
-}
