--- app/src/main/java/com/example/myriyal/utilis/AppConstants.kt
@@ -1,4 +1,0 @@
-package com.example.myriyal.utilis
-
-object AppConstants {
-}
