+++ app/src/main/java/com/example/myriyal/utils/AppConstants.kt
@@ -1,0 +1,4 @@
+package com.example.myriyal.utils
+
+object AppConstants {
+}
